+++ doc/changelog.md
@@ -15,6 +15,7 @@
 * added Green Cindarella Block (based on Green Terracotta)
 * added Soul Wood (currently unobtainable)
 * added Pucci's hat (protects from the sun)
+* added Prison Key and Plankton Vial (currently unobtainable)
 * Sinner's Block can now be crafted back into Sinner's Souls
 * Sinner's Block now activates Soul Speed and supports Soul Fire
 * Stellar Iron Ingot can now be used as beacon payment